Add handler for HikariDataSource
We have to generate a handler for each DataSource bean, but this currently only handles Hikari
Other DataSources will be done in separate PRs

On shutdown, we first suspend the pool, then request the closure of all connections and wait 30s for them all
to close.

When we restore, we simply resume the pool.

We cannot close the pool as this invalidates it, and it cannot be re-opened.

This also requires the configuration:

```
datasources.default.allow-pool-suspension: true
```

To allow pool suspension