Use the correct version of Logback to get the build working
Something is pulling in an old version of logback, which has bindings targeting slf4j-api versions 1.7.x or earlier.

Since we have moved to SLF4J 2.x, this causes an error.

I believe the cause is an old version of logback in the micronaut-bom, but until that is brought up to date, this is a
temporary workaround